Testing: Increase testing for ``LabeledDiffractionVectors2D``
